+++ src/policy/feerate.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) 2009-2010 Satoshi Nakamoto
-// Copyright (c) 2009-2015 The Bitcoin Core developers
+// Copyright (c) 2009-2018 The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "amount.h"
+#include <policy/feerate.h>
 
-#include "tinyformat.h"
+#include <tinyformat.h>
 
 const std::string CURRENCY_UNIT = "POLIS";
